[1/n] More -Wconversion fixes
The previous patches already fixes the 'big' contributors to the
-Wconversion warnings. What's left are _many_ files that each only
generate 1 or a few warnings.

Very loosely there are these broad categories:

* We have an intended narrowing conversion that also truncates the
  value. For example split a 16-bit value into 2 8-bit parts. Or
  truncating a floating point to an integer value. We now make these
  explicit via 'narrow_cast'.
* We have a narrowing conversion that's not supposed to change the
  numerical value. This is now documented (and checked in assert builds)
  via 'narrow'.
* In some cases there normally shouldn't be any change in value unless
  for wrong "user input" (e.g. a too large ROM file, or a too large
  value in a config file). These are 'fixed' by adding extra input
  validation (followed by 'narrow' or 'narrow_cast'). Though we still
  probably miss several checks.
* In some cases we have a widening (because of c/c++ language rules)
  followed by a truncation to the original width. This can fall in one
  of the previous categories (solved via 'narrow' or 'narrow_cast'). But
  sometimes it's clear from the immediate context that the truncation
  will never change the numeric value. Sometimes the compiler is smart
  enough to see this, sometimes it is not (but a newer version of the
  compiler might see it). For these I sometimes wrote a (function-style)
  cast instead of narrow/narrow_cats.

+++ src/MSXHiResTimer.cc
@@ -1,4 +1,5 @@
 #include "MSXHiResTimer.hh"
+#include "narrow.hh"
 #include "serialize.hh"
 
 namespace openmsx {
@@ -30,13 +31,13 @@ byte MSXHiResTimer::readIO(word port, EmuTime::param time)
 	}
 	// reading port {0, 1, 2, 3} will read bits {0-7, 8-15, 16-23, 24-32}
 	// of the counter
-	return latchedValue >> (8 * (port & 3));
+	return narrow_cast<byte>(latchedValue >> (8 * (port & 3)));
 }
 
 byte MSXHiResTimer::peekIO(word port, EmuTime::param time) const
 {
 	unsigned tmp = (port & 3) ? latchedValue : reference.getTicksTill(time);
-	return tmp >> (8 * (port & 3));
+	return narrow_cast<byte>(tmp >> (8 * (port & 3)));
 }
 
 template<typename Archive>
